Conflicts:
	ephub-common-rpc-connector/build.gradle

+++ ephub-product-business-service/src/main/java/com/shangpin/ephub/product/business/ui/task/spuimport/dto/HubImportTaskListRequestDto.java
@@ -21,7 +21,7 @@ public class HubImportTaskListRequestDto implements Serializable {
      * 原始文件名
      */
     private String localFileName;
-    private byte taskState;
+    private int taskState;
     private String startDate;
     private String endDate;
     private int pageNo;
